renames secureStore utils and adjusts imports
